+++ docs/changelog.md
@@ -1,6 +1,7 @@
 - Apr 20
 
     - tweaks tool bug fixes
+    - screensaver plugin
 
 - Apr 19